+++ mod/config/pkg/spec/testnet.go
@@ -23,9 +23,7 @@ package spec
 import (
 	"github.com/berachain/beacon-kit/mod/chain-spec/pkg/chain"
 	"github.com/berachain/beacon-kit/mod/primitives/pkg/common"
-	"github.com/berachain/beacon-kit/mod/primitives/pkg/crypto"
 	"github.com/berachain/beacon-kit/mod/primitives/pkg/math"
-	cmttypes "github.com/cometbft/cometbft/types"
 )
 
 // TestnetChainSpec is the ChainSpec for the bArtio testnet.
@@ -39,92 +37,4 @@ func TestnetChainSpec() (chain.Spec[
 	testnetSpec := BaseSpec()
 	testnetSpec.DepositEth1ChainID = TestnetEth1ChainID
 	return chain.NewChainSpec(testnetSpec)
-}
-
-// BaseSpec returns a chain spec with default values.
-//
-//nolint:mnd // bet.
-func BaseSpec() chain.SpecData[
-	common.DomainType,
-	math.Epoch,
-	common.ExecutionAddress,
-	math.Slot,
-	any,
-] {
-	cmtConsensusParams := cmttypes.DefaultConsensusParams()
-	cmtConsensusParams.Validator.PubKeyTypes = []string{crypto.CometBLSType}
-
-	return chain.SpecData[
-		common.DomainType,
-		math.Epoch,
-		common.ExecutionAddress,
-		math.Slot,
-		any,
-	]{
-		// Gwei value constants.
-		MinDepositAmount:          1e9,
-		MaxEffectiveBalance:       32e9,
-		EjectionBalance:           16e9,
-		EffectiveBalanceIncrement: 1e9,
-		// Time parameters constants.
-		SlotsPerEpoch:                32,
-		MinEpochsToInactivityPenalty: 4,
-		SlotsPerHistoricalRoot:       8,
-		// Signature domains.
-		DomainTypeProposer: common.DomainType{
-			0x00, 0x00, 0x00, 0x00,
-		},
-		DomainTypeAttester: common.DomainType{
-			0x01, 0x00, 0x00, 0x00,
-		},
-		DomainTypeRandao: common.DomainType{
-			0x02, 0x00, 0x00, 0x00,
-		},
-		DomainTypeDeposit: common.DomainType{
-			0x03, 0x00, 0x00, 0x00,
-		},
-		DomainTypeVoluntaryExit: common.DomainType{
-			0x04, 0x00, 0x00, 0x00,
-		},
-		DomainTypeSelectionProof: common.DomainType{
-			0x05, 0x00, 0x00, 0x00,
-		},
-		DomainTypeAggregateAndProof: common.DomainType{
-			0x06, 0x00, 0x00, 0x00,
-		},
-		DomainTypeApplicationMask: common.DomainType{
-			0x00, 0x00, 0x00, 0x01,
-		},
-		// Eth1-related values.
-		DepositContractAddress: common.NewExecutionAddressFromHex(
-			"0x4242424242424242424242424242424242424242",
-		),
-		DepositEth1ChainID:        uint64(80084),
-		Eth1FollowDistance:        1,
-		TargetSecondsPerEth1Block: 3,
-		// Fork-related values.
-		DenebPlusForkEpoch: 9999999999999998,
-		ElectraForkEpoch:   9999999999999999,
-		// State list length constants.
-		EpochsPerHistoricalVector: 8,
-		EpochsPerSlashingsVector:  8,
-		HistoricalRootsLimit:      8,
-		ValidatorRegistryLimit:    1099511627776,
-		// Max operations per block constants.
-		MaxDepositsPerBlock: 16,
-		// Slashing
-		ProportionalSlashingMultiplier: 1,
-		// Capella values.
-		MaxWithdrawalsPerPayload:         16,
-		MaxValidatorsPerWithdrawalsSweep: 1 << 14,
-		// Deneb values.
-		MinEpochsForBlobsSidecarsRequest: 4096,
-		MaxBlobCommitmentsPerBlock:       16,
-		MaxBlobsPerBlock:                 6,
-		FieldElementsPerBlob:             4096,
-		BytesPerBlob:                     131072,
-		KZGCommitmentInclusionProofDepth: 17,
-		// Comet values.
-		CometValues: cmtConsensusParams,
-	}
 }
